Update app_graphs.r
Change to whether GPUs are dropped from the Means graph's X-scale. It is now reactive based on the number of GPUs, so either show just the one GPU in the data, or all of them.
There might be circumstances this is incorrect, like working with just two GPUs, but as a general solution, I think this is best.