Supper for RData files
R has the ability to save its objects directly to files. These files can be faster to load and smaller in size than CSVs, but also will retain the R formatting, such as factors.
These changes make it possible to upload an RData file to the applet, and OCAT to RData.r will generate the RData file from a CSV and relevant TXTs. (Refer to the base OCC-OCAT scripts for how those files work.)