Diff Heat Map Toggle for Zoom
Added ability to toggle the heat map on and off for the "zoomed" Consecutive Difference graph. Not actually zoomed, but the single plot version as that should not take so long to re-render, compared to the large facet one.
By inverting things at the graph generation point, commenting out the control will not disable the heat map. The default is to including the heat map regardless of the control being in the UI.