Typo, Filtering Tweaks, and eventReactive
Tweaked the way the brush filters are done so only one logical check is made, rather than two. Should not impact performance or applet efficiency, but is the better method.
Fixed a typo (accidentally [ instead of { )
Changed the CRUSH$courseFILT object to be eventReactive instead of contained in an observeEvent. For its perhaps, this is the more correct design.