Graphics Configuration Tables
New optional section to the UI to show tables for Graphics Configuration
Currently set to show the Presets and the 60 FPS Targets configurations. May add the Options table in the future, that tracks all of the game's options, but do not feel it is necessary for Dying Light 2 at the moment.