Fix Freq Filtering bug
If you disabled a facet and double clicked on a frequency plot, it would still show the last value for the disabled facet. This was because the code to set those values to NULL was not preset.
This impacted the labels but also the filtering because the old value would continue to be used there as well.