Consecutive Difference (Percentage) Graph
Adds Consecutive Difference (Percentage) graph where the Y values and scale are actually percentages of the measurement
This, in theory, addresses the fact that the experience of a stutter is also dependent on the current frame time. (A sudden change of 4 ms means more at 10 ms than at 20 ms.) I do not think it is perfect, but this could be desirable for some.