Custom Layers and Tweaks
With various tweaks but also custom ggplot layers, GRAPH$STATS is no longer needed. One consequence of this is the zoomed QQ plot can now have the rectangles and the slope labels.
geom_qq_rect layer will draw a rectangle with the upper-right corner being the desired quantile
geom_qq_rect is applied such that the percentile range for the theoretical line adds rectangles to the graph
geom_qq_label will place a label to show the slope of the theoretical line
graphFREQ has been changed so the mean and median lines will be calculated with the graph, rather than relying on the pre-computed GRAPH$STATS
With the above changes, the zoom arguments are no longer needed for those graphs
Also, applying signif to PERCS within graphQQ addresses the 99.9% value doubling up on the scale. This occurred if the user set the QQ Percentile - High value to 99.9.