Better Graph Hiding
To prevent certain issues, the graph tabs are hidden until data is loaded. This was not working correctly before, as the tabs were always being shown , even without data. This change addresses that by making DATA$results required.
This is also done in the observeEvent that is supposed to handle updating various things when data comes in.