Fix for Frequency Stat Lines
The stat lines (mean and median) are drawn on the Frequency plot a little differently than before. With the change I forgot to have them respect the grouping set by the facet controls. This addresses that, ensuring there are not multiple sets of lines on one plot, just because the data for that plot contains more than one configuration (eg. more than one location)