Fix for Config Tabs
Realized that there was the risk of tabs constantly being added to the UI if going between datasets. My fix is to remove the tabs each time data is loaded.
Also needed to address when there is nothing for the tabs, which requires a conditional check because appendTab does not accept NULL. And, to avoid tables from prior datasets being used, those objects are removed whenever data is loaded as well.