Nested File List
The applet scans folders if it is not given a specific file to load. Now the file list it gives is grouped by the type of article the data is from (Review, Performance Analysis, Article).
The folder name is removed from this list, but a check is made in dataLOAD for the file and it will append "Data/" to the front, as that is where the file should be.