Bookmarking
Unlikely to be used, but with exploring the ability to have data to select, it seemed worth investigating.
Without data selected within the applet itself, you get the list and have a "Bookmark Data" button. The link this button gives you will load whatever the currently loaded dataset is.
Though the bookmark only gives you the dataset in the link, you can pass other input variables in the URL to apply them. (Example: adding &roundTerm=3)