Bookmarking Tweak and Note
If you provide a non-zero value for dataSelLOAD in the URL, it will immediately load the selected data. Not including this in the bookmarking button.
Unfortunately, I cannot find any immediate way to remove the data upload/selection UI elements in response to having data provided. Will continue to think on this though.